Organize Angular directory structure
