+++ site/RoomDraw/src/app/app.module.ts
@@ -4,10 +4,10 @@ import { HttpClientModule } from '@angular/common/http';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { HomeComponent } from './home/home.component';
-import { GroupComponent } from './group/group.component';
-import { NavBarComponent } from './nav-bar/nav-bar.component';
-import { HousingComponent } from './housing/housing.component';
+import { HomeComponent } from './components/home/home.component';
+import { GroupComponent } from './components/group/group.component';
+import { NavBarComponent } from './components/nav-bar/nav-bar.component';
+import { HousingComponent } from './components/housing/housing.component';
 
 @NgModule({
 	declarations: [
