+++ site/RoomDraw/src/app/components/group/group.component.spec.ts
